Move borg client class into common file
closes #8
